move filtering code to it's own module
